aberturas/requirements.txt
Django==5.1.1
django-environ==0.11.2
django-htmx==1.17.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
mysqlclient==2.2.4
redis==5.0.8
whitenoise==6.7.0
gunicorn==23.0.0
django-ratelimit==4.1.0
django-axes==6.4.0
cryptography==42.0.8
bcrypt==4.2.0